refactor: Component structure
