+++ app/src/lib/components/blocks/SocialProof.svelte
