+++ app/src/lib/components/chunks/CaseStudyCard.svelte
