+++ app/src/lib/components/chunks/StepItem.svelte
